backend/requirements.txt
# $ pip install -r requirements.txt

fastapi~=0.75.1
uvicorn~=0.17.6
requests~=2.26.0
mysql-connector-python~=8.0.27
icalendar~=4.0.9
pandas~=1.3.4
numpy~=1.22.1
